update(cgo_harness): split locked-C parity tests and refresh Go digest
- Move locked-C parity assertions behind a dedicated test file so the phase0 exclusion applies only where needed.
- Refresh the Go fixture blob digest to match the current grammar snapshot and keep deep parity checks aligned.

Buckley-Change-Hash: sha256:f46f832a79bd045956b6884a601b1c1362940654b4bba7cbd2da38348c704ccb
Buckley-Change-Stats: files=3 insertions=81 deletions=69 binaries=0
